migrated to null safety, upgraded dart sdk, tested all the changes

+++ other/kadaneAlgo.dart
@@ -13,7 +13,7 @@ int maxSubArraySum(List a, int size) {
   int curr_max = a[0];
 
   for (int i = 1; i < size; i++) {
-    curr_max = max(a[i], curr_max + a[i]);
+    curr_max = max(a[i], curr_max + a[i] as int);
     max_so_far = max(max_so_far, curr_max);
   }
   return max_so_far;
